changed some codes
