+++ backend/server.js
@@ -1,18 +1,12 @@
-import mongoose from "mongoose";
-import express from "express";
+import express, { urlencoded } from "express";
 import cors from "cors";
+import mongoose from "mongoose";
 import dotenv from "dotenv";
-import expressListEndpoints from "express-list-endpoints";
-import userRoutes from "./routes/userRoutes";
-import { connectDB } from "./config/db";
-
 dotenv.config();
+import userRoutes from "./routes/userRoutes";
 
 const mongoUrl = process.env.MONGO_URL || "mongodb://localhost/project-auth";
 mongoose.connect(mongoUrl, { useNewUrlParser: true, useUnifiedTopology: true });
-mongoose.Promise = Promise;
-
-const port = process.env.PORT || 6060;
 const app = express();
 
 app.use(cors());
@@ -20,23 +14,10 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(userRoutes);
 
-connectDB(); // Moved the connectDB function call here
-
-app.use((err, req, res, next) => {
-  const statusCode = err.statusCode || 500;
-  res.status(statusCode).json({
-    message: err.message,
-    stack: process.env.NODE_ENV === "production" ? "🥞" : err.stack,
-  });
-});
-
+// Start defining your routes here
 app.get("/", (req, res) => {
-  const endpoints = expressListEndpoints(app);
-  res.json(endpoints);
-  console.log("List of Endpoints:");
-  console.log(endpoints);
+  res.send("Hello Technigo!");
 });
-
 // Start the server
 app.listen(port, () => {
   console.log(`Server running on http://localhost:${port}`);
